test(integration): use shared `waitForNewL1Batch` implementation
