+++ core/tests/ts-integration/tests/fees.test.ts
@@ -15,7 +15,7 @@ import { TestContextOwner, TestMaster } from '../src';
 import * as zksync from 'zksync-ethers';
 import * as ethers from 'ethers';
 import { DataAvailabityMode, Token } from '../src/types';
-import { SYSTEM_CONTEXT_ADDRESS, getTestContract, waitForNewL1Batch, anyTransaction } from '../src/helpers';
+import { SYSTEM_CONTEXT_ADDRESS, getTestContract, anyTransaction } from '../src/helpers';
 import { loadConfig, shouldLoadConfigFromFile } from 'utils/build/file-configs';
 import { logsTestPath } from 'utils/build/logs';
 import { sleep } from 'utils/build';
@@ -24,6 +24,7 @@ import path from 'path';
 import { NodeSpawner } from 'utils/src/node-spawner';
 import { sendTransfers } from '../src/context-owner';
 import { Reporter } from '../src/reporter';
+import { waitForNewL1Batch } from 'utils';
 
 declare global {
     var __ZKSYNC_TEST_CONTEXT_OWNER__: TestContextOwner;
