+++ etc/utils/package.json
@@ -10,5 +10,9 @@
     },
     "dependencies": {
         "yaml": "^2.4.2"
+    },
+    "devDependencies": {
+        "ethers": "^6.13.5",
+        "zksync-ethers": "^6.18.0"
     }
 }
